doc: updated jsdocs for functions
